+++ package.json
@@ -38,6 +38,7 @@
     "events": "^3.2.0",
     "fast-sha256": "^1.3.0",
     "i18next": "^19.8.4",
+    "js-lnurl": "^0.2.5",
     "lottie-ios": "^3.1.9",
     "lottie-react-native": "^3.5.0",
     "omnibolt-js": "git+ssh://git@github.com/synonymdev/omnibolt-js",
